feat: live editor 90%
