+++ gatsby-node.js
@@ -20,6 +20,7 @@ exports.createPages = async ({ actions, graphql, reporter }) => {
   const mdxFiles = result.data.allMdx.nodes;
 
   mdxFiles.forEach(file => {
+    console.log(file.frontmatter.filterBy);
     if (file.frontmatter.type === 'module') {
       actions.createPage({
         path: `/tezos/academy/${file.frontmatter.slug}`,
@@ -28,12 +29,14 @@ exports.createPages = async ({ actions, graphql, reporter }) => {
           slug: file.frontmatter.slug,
         },
       });
-    } else if (file.frontmatter.filterBy === 'lesson-4') {
+    } else if (file.frontmatter.filterBy === 'module-04') {
+      console.log('Creating file for taquito');
       actions.createPage({
         path: `/tezos/academy/${file.frontmatter.filterBy}/${file.frontmatter.slug}`,
         component: require.resolve('./src/templates/chapterWithLiveEditor.js'),
         context: {
           slug: file.frontmatter.slug,
+          module: file.frontmatter.filterBy,
         },
       });
     } else {
