+++ src/templates/components/chapter/CodingInterfaceTaquito.js
@@ -1,16 +1,28 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect, useState, useRef } from 'react';
 import RefreshIcon from '@material-ui/icons/Refresh';
 import HelpOutlineIcon from '@material-ui/icons/HelpOutline';
 import DoneIcon from '@material-ui/icons/Done';
 import CloseIcon from '@material-ui/icons/Close';
+import styled from '@emotion/styled';
+import { ControlledEditor, monaco } from '@monaco-editor/react';
+import { TezosToolkit } from '@taquito/taquito';
+import { importKey, InMemorySigner } from '@taquito/signer';
+import StyledProvider from 'src/components/SemiLiveProvider';
+import { LiveEditor, LivePreview, LiveError } from 'react-live';
 
-import { ControlledEditor, monaco, DiffEditor } from '@monaco-editor/react';
+import theme from 'src/templates/customVSDarkTheme';
+
+import CODE_JSON from 'src/data/code';
+import STORAGE_JSON from 'src/data/storage';
+import FAUCET_KEY from 'src/data/account';
 
 const CodingInterface = ({ code, answer }) => {
   const [editorValue, setEditorValue] = useState(code);
 
   const [showAnswer, setShowAnswer] = useState(false);
   const [checkAnswer, setCheckAnswer] = useState(false);
+
+  const liveProvider = useRef(null);
 
   useEffect(() => {
     monaco.init().then(monacoInstance => {
@@ -39,48 +51,54 @@ const CodingInterface = ({ code, answer }) => {
   const resetEditorCode = () => setEditorValue(code);
 
   return (
-    <main className={`flex flex-col`}>
-      <header
-        className={`flex flex-shrink-0 justify-between items-center text-white space-x-6 h-12 bg-base-800 w-auto`}
-      >
-        <div
-          className={`bg-editor-console p-4 flex items-center justify-center h-full font-mono text-lg font-bold`}
+    <StyledProvider
+      transformCode={code => code.replace(/import .*/g, '')}
+      noInline={true}
+      ref={liveProvider}
+      //code={editorInputValue}
+
+      code={code}
+      // scope={{
+      //   ...React,
+      //   Tezos,
+      //   importKey,
+      //   InMemorySigner,
+      //   CODE_JSON,
+      //   STORAGE_JSON,
+      //   FAUCET_KEY,
+      // }}
+      // // To edit the theme, you need to edit - node_modules/prism-react-renderer/themes/vsDark/index.js
+      theme={theme}
+      style={{ overflow: 'hidden' }}
+    >
+      <main className={`flex flex-col`}>
+        <header
+          className={`flex flex-shrink-0 justify-between items-center text-white space-x-6 h-12 bg-base-800 w-auto`}
         >
-          filename.py
+          <div
+            className={`bg-editor-console p-4 flex items-center justify-center h-full font-mono text-lg font-bold`}
+          >
+            filename.py
+          </div>
+          <div className={`mr-12`}>
+            <button className={`mr-6`} onClick={resetEditorCode}>
+              <RefreshIcon />
+            </button>
+            <button className={`mr-9`}>
+              <HelpOutlineIcon />
+            </button>
+          </div>
+        </header>
+        <div className={`flex-1`}>
+          <LiveEditor
+            style={{
+              height: `calc(100vh - 14rem)`,
+              fontFamily: `"Ubuntu Mono", monospace`,
+              fontSize: '15.75px',
+              overflow: 'auto',
+            }}
+          />
         </div>
-        <div className={`mr-12`}>
-          <button className={`mr-6`} onClick={resetEditorCode}>
-            <RefreshIcon />
-          </button>
-          <button className={`mr-9`}>
-            <HelpOutlineIcon />
-          </button>
-        </div>
-      </header>
-      <div className={`flex-1`}>
-        <ControlledEditor
-          value={editorValue}
-          onChange={(_, value) => setEditorValue(value)}
-          height={`${
-            showAnswer || checkAnswer
-              ? 'calc(100vh - 8.5rem - 6rem - 15rem)'
-              : 'calc(100vh - 8.5rem - 6rem)'
-          } `}
-          language={`python`}
-          theme={`cryptoverse-theme`}
-          options={{
-            lineNumbers: true,
-            scrollBeyondLastLine: false,
-            minimap: { enabled: false },
-            scrollbar: { vertical: 'hidden', verticalScrollbarSize: 0 },
-            folding: true,
-            readOnly: false,
-            fontSize: 15.75,
-            fontFamily: "'Ubuntu Mono', monospace",
-            wordWrap: true,
-            wordBasedSuggestions: false,
-          }}
-        />
         <div
           className={`flex justify-start flex-shrink-0 bg-base-800 text-white text-lg h-12 space-x-4`}
         >
@@ -116,17 +134,19 @@ const CodingInterface = ({ code, answer }) => {
           </div>
         )}
         {showAnswer ? (
-          <DiffEditor
+          <ControlledEditor
             height={`12rem`}
-            original={showAnswer ? answer : 'MODIFIED'}
-            modified={showAnswer ? editorValue : 'MODIFIED'}
-            language="python"
-            theme={`cryptoverse-theme`}
+            value={answer}
+            language="javascript"
+            theme="cryptoverse-theme"
             options={{
               lineNumbers: false,
               scrollBeyondLastLine: false,
               minimap: { enabled: false },
-              scrollbar: { vertical: 'hidden', verticalScrollbarSize: 0 },
+              scrollbar: {
+                vertical: 'hidden',
+                verticalScrollbarSize: 0,
+              },
               folding: false,
               readOnly: true,
               fontSize: 15.75,
@@ -139,34 +159,8 @@ const CodingInterface = ({ code, answer }) => {
             }}
           />
         ) : null}
-
-        {/* <ControlledEditor
-          // height={`${
-          //   buttonClicked
-          //     ? `calc(100vh - (${HeaderHeight} + ${FooterHeight} + ${ContractFileHeight} + ${OptionHeight} + ${OutputHeaderHeight} + ${OutputContentHeight}))`
-          //     : `calc(100vh - (${HeaderHeight} + ${FooterHeight} + ${ContractFileHeight} + ${OptionHeight}))`
-          // }`}
-          // width={`calc(100vw - (100vw / 2.4))`}
-          value={editorInputValue}
-          onChange={(_, value) => {
-            setEditorInputValue(value);
-          }}
-          language="python"
-          options={{
-            lineNumbers: true,
-            scrollBeyondLastLine: false,
-            minimap: { enabled: false },
-            scrollbar: { vertical: 'hidden', verticalScrollbarSize: 0 },
-            folding: true,
-            readOnly: false,
-            fontSize: 14,
-            fontFamily: "'Ubuntu Mono', monospace",
-            wordWrap: true,
-            wordBasedSuggestions: false,
-          }}
-        /> */}
-      </div>
-    </main>
+      </main>
+    </StyledProvider>
   );
 };
